requirements.txt
streamlit
pandas
python-dateutil
python-dotenv
simplegmail
langchain-groq==0.2.4
langchain==0.3.15
markitdown
